Na, poca cosa - Soporte para varios bichos  a la vez - Algunos bugs menos

+++ src/draw/draw_attacks.cpp
@@ -2,7 +2,7 @@
 #include "defs/sprite_data.h"
 
 
-void draw_drawAttacks(TFT_eSprite &sprite, struct SpriteData* attackSpriteData, uint8_t x, uint8_t y, uint8_t attackType, uint8_t attackSprite, uint8_t factor, bool flipped) {
+void draw_drawAttacks(TFT_eSprite &sprite, struct SpriteData* attackSpriteData, int x, int y, uint8_t attackType, uint8_t attackSprite, uint8_t factor, bool flipped) {
     switch(attackType) {
         case 1:
             draw_drawSprite(sprite, x, y, attackSpriteData, attackSprite, factor, flipped);
